Rework ModuleStats to cope with repository names from Libraries.IO.
The initial version of the dependents fetch mistakenly treated the
full_name property as a package name. It is in fact a repository
name, so in order to fetch npm statistics we need to pull down and
read the package.json file in the tree and find the npm name.