Allow forcing output in CI to avoid timeouts.
Some CI systems, notably Travis, will assume the build has run off
when they observe a lack of output and the job will be killed. Try
to circumvent this by adding a flag that can force console output.

This commit also ensures that in this mode the logging output heads
to stderr such that use of the JSON reporter on CI is unaffected.