Expose the GitHub package.json request function as a static method.
This cleans up the call to the function in Liraries.IO path which
previously abused the instance to then fetch something unrelated.
This also allowed the addition of stringent tests. Do this and
also apply the .git suffix removal for consistency.

This commit also takes the opportunity to include "GitHub" in the
request creation methods where it is relevant to communicate this.