Upgrade mocha to 5.2.0.
This involved adding the "--exit" option to mocha.opts because it
seems that the timeout within installDependent() is not completely
cleaning up before it exits. While this does not affect operation
it would be nice to take heed and fix it - but this is a second
priority compared to ensuing our dependencies are more up-to-date.