Strip down the fetch command so it simply calls the shoulder CLI.
Given the support in fugl for reading package names from stdin and
the default output for shoulder >=0.4.0 being package names we no
longer need to carry the JSON array output as part of "fetch".

Pull all this code out and instead simply call the shoulder CLI.
This has the additional benefit that we inherit the warning after
breaking npm registry changes with no changes required locally.